// server/src/types/errorTypes.ts
import { ZodError } from "zod";


import ApiError from "../utils/ApiError";

export enum PrismaErrorCode {
    NOT_FOUND = "P2025",
    UNIQUE_CONSTRAINT = "P2002",
    FOREIGN_KEY = "P2003",
    VALUE_TOO_LONG = "P2000",
    VALUE_TOO_SHORT = "P2001",
    INVALID_VALUE = "P2004",
    INVALID_DATA_TYPE = "P2005",
}


export enum AppErrorCode {
    RECORD_NOT_FOUND = "RECORD_NOT_FOUND",
    DUPLICATE_ENTRY = "DUPLICATE_ENTRY",
    FOREIGN_KEY_ERROR = "FOREIGN_KEY_ERROR",
    VALIDATION_ERROR = "VALIDATION_ERROR",
    DATABASE_ERROR = "DATABASE_ERROR",
    INVALID_VALUE = "INVALID_VALUE",
    VALUE_TOO_LONG = "VALUE_TOO_LONG",
    VALUE_TOO_SHORT = "VALUE_TOO_SHORT",
    INVALID_DATA_TYPE = "INVALID_DATA_TYPE",
    API_FEATURES_FILTER_ERROR = "API_FEATURES_FILTER_ERROR",
    API_FEATURES_SORT_ERROR = "API_FEATURES_SORT_ERROR",
    API_FEATURES_FIELDS_ERROR = "API_FEATURES_FIELDS_ERROR",
    INVALID_PAGE_NUMBER = "INVALID_PAGE_NUMBER",
    INVALID_LIMIT_VALUE = "INVALID_LIMIT_VALUE",
    DATABASE_COUNT_ERROR = "DATABASE_COUNT_ERROR",
    API_FEATURES_PAGINATION_ERROR = "API_FEATURES_PAGINATION_ERROR",
    EMPTY_KEYWORD = "EMPTY_KEYWORD",
    API_FEATURES_KEYWORD_ERROR = "API_FEATURES_KEYWORD_ERROR",
    DATABASE_QUERY_ERROR = "DATABASE_QUERY_ERROR",
    API_FEATURES_EXECUTION_ERROR = "API_FEATURES_EXECUTION_ERROR",
    INVALID_MODEL_NAME = "INVALID_MODEL_NAME",
    INVALID_FIELD_NAME = "INVALID_FIELD_NAME",
    FIELD_VALIDATION_ERROR = "FIELD_VALIDATION_ERROR",
    RELATION_INCLUDE_ERROR = "RELATION_INCLUDE_ERROR",
    COMPLEX_FILTER_ERROR = "COMPLEX_FILTER_ERROR",
    INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
    INVALID_INPUT = "INVALID_INPUT",
    UNAUTHORIZED = "UNAUTHORIZED"
}


export interface ErrorResponse {
    status: string;
    message: string;
    timestamp: Date;
    errorCode?: string;
    details?: Record<string, any>;
    stack?: string;
    error?: ApiError | ZodError;
}

export interface PrismaErrorConfig {
    message: string | ((field: string) => string);
    statusCode: number;
    errorCode: AppErrorCode;
}